% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_format_climate_data.R
\name{check_format_climate_data}
\alias{check_format_climate_data}
\title{Check and format climate data to run the SWM.}
\usage{
check_format_climate_data(climate_data, rainfall_col = "pre",
  timestep = "daily", date_col = "date", date_format = "dmy",
  PET_col = NULL, light_intensity_col = NULL)
}
\arguments{
\item{climate_data}{data.frame, time series containing a date with homogen
format, and at least the corresponding rainfall data. It can also contain
PET and light intensity data, in which case the corresponding column names
have to be specified in PET_col and light_intensity_col, respectively.}

\item{rainfall_col}{single character, the name of the column containing
rainfall measurements.}

\item{timestep}{single character, defaults to "daily". If you input rainfall
data with monthly timestep (average), set this argument to "monthly".
Timesteps other than daily yields model outputs that have poorer temporal
resolution, and are not very comparable to the original model outputs
presented in Wagner et. al (2010). Using an "annual" timestep is probably
pointless, yet I included this possibility in the code.}

\item{date_col}{single character, the name of the column containing
measurement dates. Dates have to be homogenously formatted.}

\item{date_format}{single character, one of the following:
"ymd","ydm","dmy","dym","mdy","myd". y:year, m:month, d:day.}

\item{PET_col}{Optional. single character that defaults to NULL, the name of
the column containing potential evapotranspiration values to be used in the
model. If not specified, a constant, averaged daily light intensity equal
to 586.8 (from Guyaflux data) will be used as in the original model
presented by Wagner et al. (2010). This is recommended if you want to
obtain comparable result, yet the code structure technically allows to use
real measures of daily light intensity.}

\item{light_intensity_col}{Optional. single character that defaults to NULL,
the name of the column containing daily light intensity values to be used
in the model. If not specified, a constant, averaged daily light intensity
equal to 3.97 (from Guyaflux data) will be used as in the original model
presented by Wagner et al. (2010). This is recommended if you want to
obtain comparable result, yet the code structure technically allows to use
real measures of daily light intensity.}
}
\value{
A data.frame with 4 columns: date, rainfall, PET and light_intensity, with
daily timestep. If input rainfalls are monthly averages or cumulative monthly
averages, the output rainfalls will be average daily rates that are
calculated accordingly.
}
\description{
This function takes your climate data as an input, parses the date with
lubridate according to the format you specify, checks and formats the climate
date to input in the SWM. It handles timesteps other than daily, and values
of PET or light intensity different to default values of the model.
}
\details{
The function in lubridate package corresponding to the specified date format
is internally called to parse characters into Date objects. If parsing
totally of partially fails (returning NA for the corresponding lines), this
can be due to either non-homogenous date format across your dataset, or to a
wrong specification of the format. This function is not designed to tidy-up
the formats of your dataset's dates, and if any parsing fails, it will stop
with an error message. In this case, you have to check your date formats
before re-running the function.
}
