% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_format_soil_data.R
\name{check_format_soil_data}
\alias{check_format_soil_data}
\title{Format soil data for Soil Water Balance Modelling}
\usage{
check_format_soil_data(soil_data, layer_init_col = "depth_layer_init",
  layer_end_col = "depth_layer_end", FC_col = "ThetaFC",
  PWP_col = "ThetaPWP")
}
\arguments{
\item{soil_data}{data.frame, containing one-line-per-layer PWP and FC}

\item{layer_init_col}{character, the name of the column containing upper (close to the surface) end depth for each layer}

\item{layer_end_col}{character, the name of the column containing lower (far from the surface) end depth for each layer}

\item{FC_col}{character, the name of the column containing FC}

\item{PWP_col}{character, the name of the column containing PWP}
}
\description{
This function takes soil FC and PWP (measured or estimated for each layer) as an input, checks the values and returns the corresponding values at a 1cm layer-depth resolution (assuming valeus are homogenous within a measured layer)
}
