% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWB_model.R
\name{SWB_model}
\alias{SWB_model}
\title{Run the Soil Water Balance Model from Wagner et al. (2011)}
\usage{
SWB_model(climate_data, rainfall_col = "rainfall", timestep = "daily",
  date_col = "date", date_format = "ymd", PET_col = NULL,
  light_intensity_col = NULL, layer_limits_cols = c("depth_layer_init",
  "depth_layer_end"), root_dist_col = "percentRoots",
  theta_pwp_col = "ThetaPWP", theta_FC_col = "ThetaFC", ...)
}
\arguments{
\item{climate_data}{data.frame containing dates or timestamps (numbers with
separators) in a consistent format accross the table, and the corresponding
daily (or monthly, if specified) rainfall values, in milimeters.}

\item{rainfall_col}{single character, the name of the column containing
rainfall measurements.}

\item{timestep}{single character, defaults to "daily". If you input rainfall
data with monthly timestep (average), set this argument to "monthly".
Timesteps other than daily yields model outputs that have poorer temporal
resolution, and are not very comparable to the original model outputs
presented in Wagner et. al (2010). Using an "annual" timestep is probably
pointless, yet this possibility  is included in the code.}

\item{date_col}{single character, the name of the column containing
measurement dates. Dates have to be homogenously formatted.}

\item{date_format}{single character, one of the following:
"ymd","ydm","dmy","dym","mdy","myd". y:year, m:month, d:day.}

\item{PET_col}{Optional. single character that defaults to NULL, the name of
the column containing potential evapotranspiration values to be used in the
model. If not specified, a constant, averaged daily light intensity equal
to 586.8 (from Guyaflux data) will be used as in the original model
presented by Wagner et al. (2010). This is recommended if you want to
obtain comparable result, yet the code structure technically allows to use
real measures of daily light intensity.}

\item{light_intensity_col}{Optional. single character that defaults to NULL,
the name of the column containing daily light intensity values to be used
in the model. If not specified, a constant, averaged daily light intensity
equal to 3.97 (from Guyaflux data) will be used as in the original model
presented by Wagner et al. (2010). This is recommended if you want to
obtain comparable result, yet the code structure technically allows to use
real measures of daily light intensity.}

\item{...}{Any additionnal argument corresponding to model parameters, roots
or soil data tables, to replace the model's default parameters. These must
be specificated as presented in \strong{Details} section.}
}
\value{
a data.frame
}
\description{
This function runs the soil water balance model from Wagner et al. (2011)
with default parameters and soil/roots data. The user can input its own
dataset for roots sampled distribution and soil parameters (FC and PWP), or
tune any of the model parameters by adding it as additional arguments. Only
the dates and corresponding rainfall value are mandatory inputs, with
corresponding column names and timestep correctly specified.
}
\details{
Here is some additional information about the function's options and
  the formatting to input optional arguments that modify default parameters
  or datasets. Please read carefully and follow the instructions.

  \strong{Inputting different datasets for soil, roots and PAI}

  \itemize{
    \item PAI data must be inputted using argument name
    \code{PAI_data}. The content of this argument can be:
      \itemize{
        \item A scalar, the PAI value considered constant during the simulated
        timeframe
        \item A data.frame containing the same dates with the \emph{\strong{same
        format}} as those used in the climate dataset, and the corresponding PAI
        value.s, with columns named "date" and "PAI" respectively.
        \item A list containing the same elements, named "date" and "PAI".
      }

     Examples:
     \code{PAI_data = 6}
     \code{PAI_data = data.frame(climate_data$date, 6)}
     \code{PAI_data = list(date = climate_data$date, rep(6,nrow(climate_data)))}
   \item Soil data (field capacity: theta_FC and permanent wilting point: theta_PWP)
   must be inputted using argument name \code{soil_data}. The content of this argument
   can be:
      \itemize{
        \item A data.frame containing raw data to be formatted for use in the model,
        i.e. to yield FC and PWP values at a 1 cm layer depth resolution. In this case,
        the data.frame must contain the upper and lower boundaries of sampled layers,
        and the corresponding theta_FC and theta_PWP estimated values with TDR or by
        other means, with columns named "depth_layer_init", "depth_layer_end",
        "theta_FC" and "theta_PWP", respectively.
        \item A list containing the same elements, with the same requirement for names.
        \item A data.frame obtained by running \code{\link{check_format_soil_data}}
        beforehand on your raw soil data. In this case, please let the default names
        of the function's output and it will be automatically detected.
      }

     Examples:
     \code{An example with raw data as a data.frame}
     \code{An example with raw data as a list}
     \code{An example using check_format_soil_data}


   \item roots data (root distribution accross sampled soil layers) must be inputted
   using argument name \code{roots_data}. The content of this argument can be:
      \itemize{
        \item A data.frame containing raw data to be formatted for use in the model,
        i.e. to yield fine roots density values at a 1 cm layer depth resolution.
        In thiscase, the data.frame must contain the upper and lower boundaries of
        sampled layers, and the corresponding root distribution (% of total measured
        fine root quantity for all layers), with columns named "depth_layer_init",
        "depth_layer_end" and "roots_percent", respectively.
        \item A list containing the same elements, with the same requirement for names.
        \item A data.frame obtained by running \code{\link{check_format_roots_data}}
        beforehand on your raw soil data. In this case, please let the default names
        of the function's output and it will be automatically detected.
      }

     Examples:
     \code{An example with raw data as a data.frame}
     \code{An example with raw data as a list}
     \code{An example using check_format_roots_data}

     }

  \strong{Tuning any of the SWM parameters with values other than default (\emph{be careful})}

  \itemize{
    \item Transpiration parameters must be inputed using argument name
    \code{transpiration_parameters}.

    The names of the parameters and what they correspond to are:
    \itemize{
    \item \code{rho}: The ratio between transpiration and potential evapotranspiration,
    considered constant for LAI over 6 in this framework.
    \item \code{threshold}: The threshold for critical value REWc, over which water
    availability is supposed to limit tree water uptake in a given soil layer.
    }

    \code{transpiration_parameters} can be either:
      \itemize{
        \item A named numeric, that contains one or both of the parameters to be
        tuned, e.g. \code{transpiration_parameters = c(rho = 0.8, threshold = 0.4)}.
        If one of these parameters is not included in \code{transpiration_parameters},
        internal default values are used.
        \item A data.frame (1 row) containing the same parameters with appropriate
        names.
        \item A named list containing the same parameters, with appropriate names.
      }

   \item Understorey evapotranspiration parameters must be inputed using argument name
    \code{understorey_parameters}.

    The names of the parameters and what they correspond to are:
    \itemize{
    \item \code{k}: The Beer-Lambert extinction coefficient used to model light intensity
    incident to the understorey.
    \item \code{FractG}: The proportion of light incident to the understorey that is lost
    by reflection.
    \item \code{a}: Understorey evapotranspiration is hypothesized to be proportional to
    incident, non-reflected energy (light intensity). a is the proportionality
    coefficient between these two variables, in SWM's framework.
    \item \code{lambda_EU}: The coefficient of the decreasing exponential used to model
    the distribution of water uptake by understorey tree roots across the soil profile.
    It is modelled as: \code{proportion(depth) = lambda_EU*exp(-lambdaEU*abs(depth))}
    }

    \code{understorey_parameters} can be either:
      \itemize{
        \item A named numeric, that contains one or several of the parameters to be
        tuned, e.g. \code{transpiration_parameters = c(k = 0.8, a = 0.4)}.
        If one of these parameters is not included in \code{understorey_parameters},
        internal default values are used.
        \item A data.frame (1 row) containing the same parameters with appropriate
        names.
        \item A named list containing the same parameters, with appropriate names.
      }


#    \item Interception parameters must be inputed using argument name
    \code{interception_parameters}.

    The code default uses the original variant of Gash's sparse
    rainfall interception model, which is slightly different of
    what is described in Wagner et al. (2011). In total, 3
    versions of the interception function exist: The original
    code's version, the package's V1 version, and the one that
    is presented in the article. All are available in the code's
    options, and have the same input parameters.

    The names of the parameters and what they correspond to are:
    \itemize{
    \item \code{S}: the canopy compartiment's water retention capacity, or the
    amount of rainfall needed to saturate canopy under zero-evaporation conditions.
    \item \code{E_m}: the mean evaporation rate from canopies during rainfall.
    \item \code{R_m}: the mean rainfall rate (e.g. mm / minute). It is used
    to estimate the quantity of water evaporated from the canopy compartiment after
    its saturation is reached, and also the quantity of water needed to saturate
    canopy, accounting for evaporation before saturation is reached.
    \item \code{p_t}: The proportion of rain diverted to the trunk. Depending on the
    version of the submodel, this proportion applies either from the moment at which
    canopy is saturated, or for all the storm event. This is the principal difference
    between the submodel versions, and it has to be explained elsewhere later on.
    \item \code{S_t}: The trunks' water retention capacity
    \item \code{cc}: The percentage of canopy cover (canopy surface area per ground
    area).
    }

    \code{interception_parameters} can be either:
      \itemize{
        \item A named numeric, that contains one or several of the parameters to be
        tuned, e.g. \code{transpiration_parameters = c(S = 1.8, E_m = 6.4)}.
        If one of these parameters is not included in \code{understorey_parameters},
        internal default values are used.
        \item A data.frame (1 row) containing the same parameters with appropriate
        names.
        \item A named list containing the same parameters, with appropriate names.
      }




  Nino out }

  Then another sub section
}
